pyproject.toml
[project]
name = "ai-video-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "pydantic-ai>=1.1.0",
    "pydantic-settings>=2.11.0",
    "whisperx==3.3.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.0",
]

[tool.ruff]
exclude = [
    ".git",
    ".vscode",
    ".ruff_cache",
    ".venv",
    "venv",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"